temporary fix for moveTo and negative coordinates
The problem with removing the >0 restriction on moveToX:Y:Z is that now
all the instances where we are passing in -1 for “don’t move on this
axis” cause problems during a scan (maybe because this causes too many
commands in rapid succession, not to mention it slightly moves the
steppers when we don’t want). So I’ve added an exception only for -1.
But would be better to have BOOLs in here for each axis. I’ll think on
this a bit.